Filter buffer size settings cleanup

+++ sonification-workstation/SonificationWorkstation/smafilter.h
@@ -13,7 +13,7 @@ public:
     explicit SmaFilter(QObject *parent = nullptr);
     virtual void push(float value) override;
     virtual float value() override;
-    virtual void setN(float n) override;
+    virtual void setN(uint n) override;
     virtual void flush() override;
 
 private:
